api/requirements.txt
# Core Django framework
# Specify a range to ensure compatibility with future minor updates while avoiding potential breaking changes in new major versions.
Django>=5.0.2,<5.1

# ASGI interface library for async capability with Django
# Allow minor updates for fixes and improvements.
asgiref>=3.7.2,<3.8

# WSGI HTTP Server for UNIX to serve Django projects
# Allow minor updates for stability and performance improvements.
gunicorn>=21.2.0,<21.3

# Package for version handling, allow minor updates.
packaging>=23.2,<23.3

# PostgreSQL adapter for Django
# Specify a range to maintain compatibility with newer versions while avoiding breaking changes.
psycopg2-binary>=2.9.9,<2.10

# Parser for SQL statements, useful for Django's ORM
# Minor updates can offer parsing improvements.
sqlparse>=0.4.4,<0.5

# Django REST framework for building APIs
# Pin to a specific major version for API stability, while allowing minor updates within that major version.
djangorestframework>=3.14.0,<3.15

# Library for making HTTP requests, allow minor updates for improvements and security patches.
requests>=2.31.0,<2.32

# OpenAPI schema generation for DRF, minor updates for enhancements and fixes.
drf-spectacular>=0.27.1,<0.28

# JWT handling for authentication, support minor updates.
PyJWT>=2.8.0,<2.9

cryptography>=42.0.2,<43

requests_oauthlib>=1.3.1
